Product page - image easy mode
Set up the product page with the imageUrl - not sure how to save images as blobs on the database.
Tried to write a uploadImage.jsp ... it writes .. something(?) to the db.
It loads an image and reads it into a byte array, then I try to convert it to a blob using some method I found online. Not sure if it works. Ill try and implement the dispalyImage.jsp tmr and see if it actually works.